unit work

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_modules/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   pageNumber = 1;
   pageSize = 5;
   loading = false;
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmServic: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,10 +33,17 @@ export class MessagesComponent implements OnInit {
     })
   }
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => { this.messages.splice(this.messages.findIndex(m => m.id === id, 1)) }
+    this.confirmServic.confirm('Confirm delete Message', 'this cannot be undone').subscribe({
+      next: result => {
+        if (result) {
+          this.messageService.deleteMessage(id).subscribe({
+            next: () => { this.messages.splice(this.messages.findIndex(m => m.id === id),1) }
 
+          })
+        }
+      }
     })
+
   }
   pageChanged(event: any) {
     if (this.pageNumber !== event.page) {
